Filter out SEPA when PMs are reloaded
