+++ paymentsheet/src/main/java/com/stripe/android/customersheet/data/CustomerSessionPaymentMethodDataSource.kt
@@ -1,6 +1,8 @@
 package com.stripe.android.customersheet.data
 
 import com.stripe.android.core.injection.IOContext
+import com.stripe.android.customersheet.util.filterToSupportedPaymentMethods
+import com.stripe.android.customersheet.util.getDefaultPaymentMethodsEnabledForCustomerSheet
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.PaymentMethodUpdateParams
 import com.stripe.android.payments.core.analytics.ErrorReporter
@@ -18,7 +20,13 @@ internal class CustomerSessionPaymentMethodDataSource @Inject constructor(
     override suspend fun retrievePaymentMethods(): CustomerSheetDataResult<List<PaymentMethod>> {
         return withContext(workContext) {
             elementsSessionManager.fetchElementsSession().mapCatching { elementsSessionWithCustomer ->
-                elementsSessionWithCustomer.customer.paymentMethods
+                val isSyncDefaultPaymentMethodFeatureEnabled = getDefaultPaymentMethodsEnabledForCustomerSheet(
+                    elementsSessionWithCustomer.elementsSession
+                )
+
+                elementsSessionWithCustomer.customer.paymentMethods.filterToSupportedPaymentMethods(
+                    isSyncDefaultPaymentMethodFeatureEnabled
+                )
             }.toCustomerSheetDataResult()
         }
     }
